replace CPGraph by Array as state (CPGraph is still used)
Using a different type than Array make a lot of stuffs more difficult and make RL.jl partly unusable. As it is quite easy to get from CPGraph to an Array of dimension 2 and vice versa, we choose to encode and decode the CPGraph. We also delete CPGraphSpace, which, in any case, is not useful