RL pipeline refactoring and GPU support (#158)
Major SeaPearl update !

The main changes are related to the ReinforcementLearning pipeline. Our goal was to get rid of the ArrayBufferStateRepresentation that we considered inefficient and too restrictive. We instead provided SeaPearl with a new kind of states: the AbstractTrajectoryState which could be understood as a bridge between the inner AbstractStateRepresentation and the RL agent. We provided by default a DefaultTrajectoryState which only consists in a FeaturedGraph and the index of the variable we are branching on. This DefaultTrajectoryState is thus a lighter version of the inner state representation, with no restriction whatsoever.

We also added a BatchedDefaultTrajectoryState, which contains all the information of a batch of DefaultTrajectoryState but with data formatted for efficient parallel computation (i.e. as 3D arrays). The size of these arrays is determined dynamically when sending the batch to the target device and creating the BatchedDefaultTrajectoryState (though the functor function).

The novelties are:

    new struct DefaultTrajectoryState;
    new struct BatchedDefaultTrajectoryState;
    obligation to provide an AbstractTrajectoryState for the StateRepresentation;
    a GeometricFlux.GraphConv dispatch for BatchedDefaultTrajectoryState (impossible to do parallel computation with the existing code of GeometricFlux);
    revision of FlexGNN for the new input type;
    easy conversion of CPLayerGraph to adjacency matrix using LigthGraphs;
    new subtypes of AbstractTrajectoryState (Tabular and NonTabular) to specify the behavior of RL.jl;
    possibility to update the features at each step of the search;
    GPU support for the GNN computation;
    existing code refactoring;
    some CP bugs were found and corrected;


Co-authored-by: Tom <tom.marty@polytechnique.edu>
Co-authored-by: tomsander1998 <tom.sander@polytechnique.edu>