fix CPModification update with View Variables (#160)
As observed almost a year ago in #55 , CPModification isn't updated correctly with IntVarView and BoolVarView.
I propose a fix by providing a new definition of addToPrunedDomains for VariableViews. After creating/updating the entry corresponding to the view, it recursively call the function for the parent variable until a root variable is found.