removed mean computation (#175)
In the basic metrics, a mean computation is integrated. But this mean uses arbitrary parameters and prevent the possibility to store raw-data in the final csv file. In my opinion this kind of operation should be done in post-treatment, when plotting data, rather than during data-gathering. (this function makes it impossible to use the storedata function from SeaPearlExtras)